Add keypress event listener for player selection
Adds a keypress event listener listening for a and l, to determine whether player 1 or player 2 will select cards that make a set respectively.
Adds a container with instructions on these controls.
Adds a pick card reminder that says which player is picking and how many more cards they have to select.